requirements.txt
python-dotenv
redis
Flask
apscheduler
imaplib2